// src/app/Components/booking/Model/Update-booking.request.model.ts
export interface UpdateBookingRequest{

    Pickup: string;
    Via: string;
    DropOff: string;
    PickUpDate: string;
    Passengers: string;
    Louggages: string;
    Greet: boolean;
    Price: number;
    CarType: string;
    CarImage:string;
    Name: string;
    PhoneNumber: string;
    Email: string;
    DriverInstruction: string;
}